sope-mime: remove encoding of values of MIME headers
The UTF-8 'Q' encoding makes Outlook to not recognize attachments.
Also Thunderbird loses the attached file names.

Removing this encoding and sending directly UTF-8 data make
attachments work in Outlook, Webmail and Thunderbird.

This does not seem a correct by standards solution and may trigger
problems with other clients or servers which don't have the SMTPUTF8
extension.
However this allow us to work correctly with our main targets.

Note that header values in the mail main part continue to be
encoded and that has not presented any problem so far.