Use same way to get NSStringEncoding in similar methods
Currently there are two methods to get the NSStringEncoding
from a named encoding [NGMime stringEncodingForCharset:] and
[NSString stringEncodingForEncodingNamed:] this two are not
only redundant but use a different way to match the NSStringEncoding
with its name. This can spell difficult to trace bugs.

[NGMime stringEncodingForCharset:] uses a hardcoded table.

[NSString stringEncodingForEncodingNamed:] depends on the system,
compiled with GNUStep uses a call to GSMimeDocument.

This commit makes [NGMime stringEncodingForCharset:] to call
[NSString stringEncodingForEncodingNamed:] to assure we use
the same resolution method.

This does not make the methods totally equivalent, with unknown
encodings [NSString stringEncodingForEncodingNamed:] returns 0
but [NGMime stringEncodingForCharset:] returns either
NSISOLatin1StringEncoding or [NSString defaultCStringEncoding].
The last value is returned for nil or empty charset names.

If this changset is accepted, the next step would be to replace
calls to [NGMime stringEncodingForCharset:] with
[NString stringEncodingForEncodingNamed:] calls.